+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   - Added `target` option to JDBC input, allowing the row columns to target a specific field instead of being expanded 
     at the root of the event. This allows the input to play nicer with the Elastic Common Schema when 
     the input does not follow the schema. [#69](https://github.com/logstash-plugins/logstash-integration-jdbc/issues/69)
+    
+  - Added `target` to JDBC filter static `local_lookups` to verify it's properly valued when ECS is enabled. 
+    [#71](https://github.com/logstash-plugins/logstash-integration-jdbc/issues/71)
 
 ## 5.0.7
   - Feat: try hard to log Java cause (chain) [#62](https://github.com/logstash-plugins/logstash-integration-jdbc/pull/62)